requirements.txt
streamlit
tensorflow
opencv-python
pillow
matplotlib
numpy